Paragraph: avoid re-rendering block contols on type
