+++ packages/block-library/src/paragraph/controls.js
@@ -1,0 +1,110 @@
+/**
+ * WordPress dependencies
+ */
+import { __, _x, isRTL } from '@wordpress/i18n';
+import {
+	ToolbarButton,
+	ToggleControl,
+	__experimentalToolsPanelItem as ToolsPanelItem,
+} from '@wordpress/components';
+import {
+	AlignmentControl,
+	BlockControls,
+	InspectorControls,
+	useSettings,
+	store as blockEditorStore,
+} from '@wordpress/block-editor';
+import { formatLtr } from '@wordpress/icons';
+import { useSelect } from '@wordpress/data';
+import { pure } from '@wordpress/compose';
+
+function ParagraphRTLControl( { direction, setDirection } ) {
+	return (
+		isRTL() && (
+			<ToolbarButton
+				icon={ formatLtr }
+				title={ _x( 'Left to right', 'editor button' ) }
+				isActive={ direction === 'ltr' }
+				onClick={ () => {
+					setDirection( direction === 'ltr' ? undefined : 'ltr' );
+				} }
+			/>
+		)
+	);
+}
+
+export function hasDropCapDisabled( align ) {
+	return align === ( isRTL() ? 'left' : 'right' ) || align === 'center';
+}
+
+function Controls( { clientId, setAttributes } ) {
+	const [ isDropCapFeatureEnabled ] = useSettings( 'typography.dropCap' );
+
+	function selector( select ) {
+		const { align, direction, dropCap } =
+			select( blockEditorStore ).getBlockAttributes( clientId ) || {};
+		return { align, direction, dropCap };
+	}
+	const { align, direction, dropCap } = useSelect( selector, [ clientId ] );
+
+	let helpText;
+	if ( hasDropCapDisabled( align ) ) {
+		helpText = __( 'Not available for aligned text.' );
+	} else if ( dropCap ) {
+		helpText = __( 'Showing large initial letter.' );
+	} else {
+		helpText = __( 'Toggle to show a large initial letter.' );
+	}
+
+	return (
+		<>
+			<BlockControls group="block">
+				<AlignmentControl
+					value={ align }
+					onChange={ ( newAlign ) =>
+						setAttributes( {
+							align: newAlign,
+							dropCap: hasDropCapDisabled( newAlign )
+								? false
+								: dropCap,
+						} )
+					}
+				/>
+				<ParagraphRTLControl
+					direction={ direction }
+					setDirection={ ( newDirection ) =>
+						setAttributes( { direction: newDirection } )
+					}
+				/>
+			</BlockControls>
+			{ isDropCapFeatureEnabled && (
+				<InspectorControls group="typography">
+					<ToolsPanelItem
+						hasValue={ () => !! dropCap }
+						label={ __( 'Drop cap' ) }
+						onDeselect={ () =>
+							setAttributes( { dropCap: undefined } )
+						}
+						resetAllFilter={ () => ( { dropCap: undefined } ) }
+						panelId={ clientId }
+					>
+						<ToggleControl
+							__nextHasNoMarginBottom
+							label={ __( 'Drop cap' ) }
+							checked={ !! dropCap }
+							onChange={ () =>
+								setAttributes( { dropCap: ! dropCap } )
+							}
+							help={ helpText }
+							disabled={
+								hasDropCapDisabled( align ) ? true : false
+							}
+						/>
+					</ToolsPanelItem>
+				</InspectorControls>
+			) }
+		</>
+	);
+}
+
+export default pure( Controls );
